+++ src/lossFunctions/loss.hpp
@@ -3,9 +3,9 @@
 
 class loss {
 public:
-	virtual nc::NdArray<double> forward(const nc::NdArray<double>&, const nc::NdArray<int>&) = 0;
+	virtual nc::NdArray<double> forward(const nc::NdArray<double>&, const nc::NdArray<size_t>&) = 0;
 
-	nc::NdArray<double> calculate(const nc::NdArray<double>& output, const nc::NdArray<int>& y)
+	nc::NdArray<double> calculate(const nc::NdArray<double>& output, const nc::NdArray<size_t>& y)
 	{
 		//Calculates the average loss
 		return nc::mean(this->forward(output, y));
